Imporve the create tenant command
