+++ app/Multitenancy/Events/NewTenantSetupComplete.php
@@ -1,21 +1,17 @@
 <?php
 
-namespace App\Events;
+namespace App\Multitenancy\Events;
 
-use Throwable;
 use App\Models\Tenant;
 use Illuminate\Queue\SerializesModels;
 use Illuminate\Foundation\Events\Dispatchable;
 use Illuminate\Broadcasting\InteractsWithSockets;
 
-class NewTenantSetupFailure
+class NewTenantSetupComplete
 {
     use Dispatchable;
     use InteractsWithSockets;
     use SerializesModels;
 
-    public function __construct(
-        public Tenant $tenant,
-        public Throwable $exception,
-    ) {}
+    public function __construct(public Tenant $tenant) {}
 }
