+++ app/Multitenancy/Events/NewTenantSetupFailure.php
@@ -1,17 +1,21 @@
 <?php
 
-namespace App\Events;
+namespace App\Multitenancy\Events;
 
+use Throwable;
 use App\Models\Tenant;
 use Illuminate\Queue\SerializesModels;
 use Illuminate\Foundation\Events\Dispatchable;
 use Illuminate\Broadcasting\InteractsWithSockets;
 
-class NewTenantSetupComplete
+class NewTenantSetupFailure
 {
     use Dispatchable;
     use InteractsWithSockets;
     use SerializesModels;
 
-    public function __construct(public Tenant $tenant) {}
+    public function __construct(
+        public Tenant $tenant,
+        public Throwable $exception,
+    ) {}
 }
