+++ app/Observers/TenantObserver.php
@@ -8,11 +8,11 @@ use Illuminate\Bus\Batch;
 use App\Jobs\MigrateTenantDatabase;
 use App\Jobs\SetupMaterializedView;
 use Illuminate\Support\Facades\Bus;
-use App\Events\NewTenantSetupFailure;
 use App\Jobs\SetupForeignDataWrapper;
 use Illuminate\Support\Facades\Event;
-use App\Events\NewTenantSetupComplete;
 use App\Jobs\MigrateTenantSisDatabase;
+use App\Multitenancy\Events\NewTenantSetupFailure;
+use App\Multitenancy\Events\NewTenantSetupComplete;
 
 class TenantObserver
 {
